@@ -1,7 +1,7 @@
 {
     "name": "willdurand/email-reply-parser",
     "type": "library",
-    "description": "Port of the cool GitHub's EmailReplyParser library in PHP 5.3",
+    "description": "Port of the cool GitHub's EmailReplyParser library in PHP",
     "keywords": ["email", "reply-parser"],
     "license": "MIT",
     "authors": [
@@ -11,7 +11,7 @@
         }
     ],
     "require": {
-        "php": ">=5.3.0"
+        "php": ">=5.6.0"
     },
     "autoload": {
         "psr-4": { "EmailReplyParser\\": "src/EmailReplyParser" }
@@ -21,7 +21,10 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "2.5-dev"
+            "dev-master": "2.6-dev"
         }
+    },
+    "require-dev": {
+        "phpunit/phpunit": "^4.8.35|^5.7"
     }
 }
