@@ -2,7 +2,7 @@
 
 namespace EmailReplyParser\Tests;
 
-class TestCase extends \PHPUnit_Framework_TestCase
+abstract class TestCase extends \PHPUnit\Framework\TestCase
 {
     /**
      * @param string $file
